Added flush and isBefore methods to ROB.

+++ simulator/src/simulator/Instruction.java
@@ -15,7 +15,7 @@ public class Instruction {
 	
 	public class Operand {
 		Integer intValue;
-        Double floatValue;
+        Float floatValue;
 		Integer registerIndex;
 		RegisterType registerType;
